machine-learning/cloudrun/requirements.txt
Flask==1.1.2
pytest==5.3.0; python_version > "3.0"
pytest==4.6.6; python_version < "3.0"
gunicorn==20.0.4
google-cloud-automl==1.0.1
google-cloud-storage==1.29.0